nvim: cleanup lspconfig for new version of neovim
- mason-lspconfig doesn't provide a function for initializing language
servers anymore since neovim can do it. As such, migrate to using a lua
file for each ls to configure it
- replace folke/neodev with folke/lazydev (better version of the same
thing)
- move much of what was previously being jankily called in
mason-lspconfig's setup into lsp.lua, and call that from init.lua
- clean up what was left of lspconfig.lua after the above
- update lazy lockfile
- remove explicit call to clangd_extensions `setup` since what I was
passing to it wasn't even doing anything. leave everything default
instead
