+++ .config/nvim/after/lsp/basedpyright.lua
@@ -1,0 +1,23 @@
+return {
+  on_attach = function(client, bufnr)
+    vim.api.nvim_buf_create_user_command(bufnr, "LspPyrightOrganizeImports", function()
+      client:exec_cmd({
+        command = "basedpyright.organizeimports",
+        arguments = { vim.uri_from_bufnr(bufnr) },
+      })
+    end, {
+      desc = "Organize Imports",
+    })
+    vim.g.my_on_attach(client, bufnr)
+  end,
+
+  settings = {
+    basedpyright = {
+      analysis = {
+        autoSearchPaths = true,
+        diagnosticMode = "openFilesOnly",
+        useLibraryCodeForTypes = true,
+      },
+    },
+  },
+}
