+++ .config/nvim/after/lsp/nil_ls.lua
@@ -1,0 +1,11 @@
+return {
+  settings = {
+    ["nil"] = {
+      nix = {
+        flake = {
+          autoArchive = true,
+        },
+      },
+    },
+  },
+}
