+++ .config/nvim/after/lsp/rust_analyzer.lua
@@ -1,0 +1,45 @@
+return {
+  cmd = { vim.fn.executable("ra-multiplex") and "ra-multiplex" or "rust-analyzer" },
+  settings = {
+    ["rust-analyzer"] = {
+      files = {
+        watcher = "server",
+      },
+      check = {
+        workspace = false,
+      },
+      inlayHints = {
+        bindingModeHints = {
+          enable = false,
+        },
+        chainingHints = {
+          enable = true,
+        },
+        closingBraceHints = {
+          enable = true,
+          minLines = 25,
+        },
+        closureReturnTypeHints = {
+          enable = "never",
+        },
+        lifetimeElisionHints = {
+          enable = "never",
+          useParameterNames = false,
+        },
+        maxLength = 25,
+        parameterHints = {
+          enable = true,
+        },
+        reborrowHints = {
+          enable = "never",
+        },
+        renderColons = true,
+        typeHints = {
+          enable = true,
+          hideClosureInitialization = false,
+          hideNamedConstructor = false,
+        },
+      },
+    },
+  },
+}
