+++ .config/nvim/after/lsp/yamlls.lua
@@ -1,0 +1,7 @@
+return {
+  settings = {
+    yaml = {
+      keyOrdering = false,
+    },
+  },
+}
